rest/requirements.txt
click==6.6
ConcurrentLogHandler==0.9.1
Flask==0.11
funcsigs==1.0.2
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
kafka-python==1.3.1
kazoo==2.2.1
MarkupSafe==0.23
mock==2.0.0
nose==1.3.7
pbr==1.10.0
python-json-logger==0.1.5
redis==2.10.5
requests==2.11.1
retrying==1.3.3
scutils==1.1.0
six==1.10.0
testfixtures==4.11.0
ujson==1.35
Werkzeug==0.11.11